# Add support for CompletionList.Data in LSP.