--- a/src/Features/LanguageServer/ProtocolUnitTests/Completion/CompletionResolveTests.cs
+++ b/src/Features/LanguageServer/ProtocolUnitTests/Completion/CompletionResolveTests.cs
@@ -273,7 +273,8 @@ class B : A
             {
                 foreach (var completionItem in completionList.Items)
                 {
-                    completionItem.Data ??= vsCompletionList.Data;
+                    Assert.Null(completionItem.Data);
+                    completionItem.Data = vsCompletionList.Data;
                 }
             }