- Ultimately this is a performance optimization where instead of re-specifying resolve data on every completion item the platform now can specify if it supports specifying resolve data at the completion list level.
- Added new completion resolve test and updated testing infrastructure to lift resolve data from completion list.

**Before:**
|                                                    Method | Optimized |      Mean |     Error |    StdDev |   Op/s |     Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|---------------------------------------------------------- |---------- |----------:|----------:|----------:|-------:|----------:|---------:|---------:|----------:|
| 'Completion List Roundtrip Serialization/Deserialization' |     False | 29.379 ms | 0.5844 ms | 1.0976 ms |  34.04 | 1562.5000 | 593.7500 | 250.0000 |   9.48 MB |
|                           'Completion List Serialization' |     False | 11.038 ms | 0.2194 ms | 0.2155 ms |  90.60 |  562.5000 | 281.2500 | 281.2500 |   3.33 MB |
|                         'Completion List Deserialization' |     False | 15.353 ms | 0.2995 ms | 0.2941 ms |  65.13 |  921.8750 | 375.0000 |        - |   5.57 MB |
| 'Completion List Roundtrip Serialization/Deserialization' |      True | 18.469 ms | 0.1742 ms | 0.1455 ms |  54.15 | 1093.7500 | 500.0000 | 125.0000 |   7.09 MB |
|                           'Completion List Serialization' |      True |  4.574 ms | 0.0823 ms | 0.1207 ms | 218.64 |  242.1875 | 117.1875 | 117.1875 |   1.53 MB |
|                         'Completion List Deserialization' |      True | 15.000 ms | 0.0846 ms | 0.0707 ms |  66.67 |  921.8750 | 375.0000 |        - |   5.57 MB |

**After:**
|                                                    Method | Optimized |      Mean |     Error |    StdDev |   Op/s |     Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|---------------------------------------------------------- |---------- |----------:|----------:|----------:|-------:|----------:|---------:|---------:|----------:|
| 'Completion List Roundtrip Serialization/Deserialization' |     False | 19.500 ms | 0.2663 ms | 0.2224 ms |  51.28 | 1218.7500 | 406.2500 | 125.0000 |   7.71 MB |
|                           'Completion List Serialization' |     False |  8.026 ms | 0.1524 ms | 0.1351 ms | 124.60 |  406.2500 | 171.8750 | 109.3750 |    2.7 MB |
|                         'Completion List Deserialization' |     False | 10.513 ms | 0.2078 ms | 0.2845 ms |  95.12 |  734.3750 | 250.0000 |        - |   4.42 MB |
| 'Completion List Roundtrip Serialization/Deserialization' |      True | 12.622 ms | 0.2425 ms | 0.3153 ms |  79.22 |  906.2500 | 265.6250 | 125.0000 |   5.82 MB |
|                           'Completion List Serialization' |      True |  2.043 ms | 0.0089 ms | 0.0079 ms | 489.44 |  246.0938 | 121.0938 | 121.0938 |    1.4 MB |
|                         'Completion List Deserialization' |      True | 10.563 ms | 0.2048 ms | 0.2734 ms |  94.67 |  734.3750 | 250.0000 |        - |   4.42 MB |

**Results (Optimized):**
Roundtrip: 1.5x faster
Serialization: 2.2x faster
Deserialization: 1.4x faster

Fixes dotnet/aspnetcore#31883